table pagination refactoring

+++ app/Traits/Paginatable.php
@@ -9,8 +9,8 @@ trait Paginatable
     {
         $pageSize = request('page_size', config('custom.system.pagination_items_per_page'));
         if ($pageSize === 'all') {
-            return config('custom.system.pagination_max_limit');
+            return (int)config('custom.system.pagination_max_limit');
         }
-        return $pageSize;
+        return (int)$pageSize;
     }
 }
